// src/app/models/ims/inventory-product-history.ts
export class InventoryProductHistory {
    barCode: string;
    presentation: string;
    productName: string;
    scaleCode: string;
    area: string;
    initialInventory: number;
    sales: number;
    salesReturn: number;
    transfers: number;
    adjustments: number;
    purchases: number;
    purchaseReturns: number;
    productions: number;
    packing: number;
    unPacking: number;
    finalInventory: number;
    category: string;
    hierarchyCategory: string;
    supplier: string;
    brand: string;
    productEstatus: string;
}